requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.5
django-cors-headers==4.7.0
<<<<<<< HEAD
django-filter==25.2
=======
>>>>>>> 6de585d (Added modules from Module 2 and 3. Added Faculty Feedback Message, Final Grade Added Admin View.)
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.5
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
